BBBSWA-20 Added defaults map. Component working with router
