+++ frontEnd/src/assets/defaults.json
@@ -1,0 +1,74 @@
+{
+  "BIG Profile": {
+    "description": "Congratulations on making your account!",
+    "upload": false
+  },
+  "BIG Chat": {
+    "description": "A quick call with a BBBS team member to answer any questions, discuss program options and help keep your journey moving.",
+    "upload": false
+  },
+  "Mentoring 101": {
+    "description": "An online orientation from the Alberta Mentoring Partnership that prepares you to start your journey as a mentor.  This online training will take about an hour. http://moodle.albertamentors.ca",
+    "upload": false
+  },
+  "BIG Supporters - Family/Partner": {
+    "description": "This reference must have known your for a minimum of two years. Please provide significant other, if not applicable, please provide family member.",
+    "upload": false
+  },
+  "BIG Supporters - Personal": {
+    "description": "This reference must have known your for a minimum of two years.",
+    "upload": false
+  },
+  "BIG Supporters - Employer": {
+    "description": "This reference must have known your for a minimum of two years. Please provide your current employer, if this individual has known you less than two years, please be prepared to provide an additional reference.",
+    "upload": false
+  },
+  "BIG Fundamentals": {
+    "description": "Your 3.5 hour training session with BBBS staff and other mentors. Flexiblity for training dates will be made avaliable through our scheduling database - Acuity. A link to Acuity will be sent to you by BBBS upon completion of your prescreen",
+    "upload": false
+  },
+  "BIG Checks - Police": {
+    "description": "When you have received your results from the Calgary Police, please share the results throught the Calgary Police Website to BBBS!",
+    "upload": false
+  },
+  "BIG Checks - RCMP Check": {
+    "description": "Please remember to share with BBBS! If you live out of the city of Calgary, be prepared that you will be completing a RCMP check instead.",
+    "upload": false
+  },
+  "BIG Checks - Intervention Record Check": {
+    "description": "Will be provided to you electronically or in paper form from BBBS.",
+    "upload": false
+  },
+  "BIG Bio": {
+    "description": "Your interview with our BBBS team to get to know you and help determine your matchablity.",
+    "upload": false
+  },
+  "BIG Extras - Confidentiality": {
+    "description": "This document is to ensure privacy for all parties involved.",
+    "upload": false
+  },
+  "BIG Extras - Media Consent": {
+    "description": "This document gives the option for you image to be used in BBBS communication and marketing materials.",
+    "upload": false
+  },
+  "BIG Extras - Offence Declaration": {
+    "description": "An additional and ongoing form to discuss to declare any changes in your criminal record status.",
+    "upload": true
+  },
+  "BIG Extras - Home Assessment": {
+    "description": "If deemed necessary at interview",
+    "upload": false
+  },
+  "BIG Extras - Car Insurance": {
+    "description": "Please provide a copy of your full page policy (not just the pink slip), indicating your Third Party Liability coverage as at least $1,000,000",
+    "upload": false
+  },
+  "You are a BIG Deal!": {
+    "description": "You are ready to be matched with your Little Brother/Sister!",
+    "upload": false
+  },
+  "You are no BIG Deal :(": {
+    "description": "Did not get accepted",
+    "upload": false
+  }
+}
